+++ src/dnssec/utils/legacy/privkey.c
@@ -87,7 +87,7 @@ const param_t PRIVKEY_CONVERSION_TABLE[] = {
 static int parse_algorithm(char *string, void *_algorithm)
 {
 	char *end = string;
-	while (*end != '\0' && !isspace(*end)) {
+	while (*end != '\0' && !isspace((int)*end)) {
 		end += 1;
 	}
 	*end = '\0';
@@ -148,12 +148,12 @@ static int parse_time(char *string, void *_time)
 static void strip(char **value, size_t *length)
 {
 	// strip from left
-	while (*length > 0 && isspace(**value)) {
+	while (*length > 0 && isspace((int)**value)) {
 		*value += 1;
 		*length -= 1;
 	}
 	// strip from right
-	while (*length > 0 && isspace((*value)[*length - 1])) {
+	while (*length > 0 && isspace((int)(*value)[*length - 1])) {
 		*length -= 1;
 	}
 }
